fix return type of the PyUnicode_KIND macro
The documentation of the PyUnicode_KIND macro specifies a return value of type int.
However the macro currently returns a value of the underlying type of the enum PyUnicode_Kind, which is implementation defined.

+++ Include/cpython/unicodeobject.h
@@ -334,7 +334,7 @@ enum PyUnicode_Kind {
 #define PyUnicode_KIND(op) \
     (assert(PyUnicode_Check(op)), \
      assert(PyUnicode_IS_READY(op)),            \
-     ((PyASCIIObject *)(op))->state.kind)
+     ((int)((PyASCIIObject *)(op))->state.kind))
 
 /* Return a void pointer to the raw unicode buffer. */
 #define _PyUnicode_COMPACT_DATA(op)                     \
